requirements.txt
# pip install -r requirements.txt
crewai
exa_py
langchain
python-dotenv
streamlit
groq
